A FakerProvider for code bases with a PSR-11 compliant DI container (#1)
* FakerProvider that pulls the Generator from a PSR-11 container.

As it stands, this code only supports pulling the container itself from a PHP file. That file may
either return the container object directly, or set it into a defined variable name. The provider
will include the PHP file defined to get the container, then return the Generator from the container
by using the defined ID.

By using the Generator from the DI container we ensure that all custom providers are added to the
Generator first and so making them available for analysis.

As an example... if using Symfony, something similar to:
    $provider = new PsrContainerFakerProvider(
        './var/cache/dev/App_KernelDevDebugContainer.php',
        null,
        'Faker\Generator'
    );

will give us a provider that, when asked for the Faker Generator, will include the
App_KernelDevDebugContainer.php file, using the return value as the container. From that container
the provider will then pull a service with the ID 'Faker\Generator' and return it.

* A factory that can provide a PsrContainerFakerProvider.

To allow for the service that provides \CalebDW\Fakerstan\FakerProvider to call the
static 'create' method on the defined factory, this factory takes parameters for its
constructor and then copies those values onto static member variables. The static
'create' method then has access to the values that were given to the constructor.

* Add information about PsrContainerFakerProviderFactory to documentation.

* Linting.

* Import classes instead of using FQCNs.

* Call TestCase methods on $this, instead of self.

* Use it* instead of getFaker*.

* Remove unnecessary comments.

* Just require the container file, instead of using require_once.

This removes the issue of the changing return value when calling require_once multiple
times. From that, it makes our tests easier to set up and tear down, too.

* Use the arguments key in example definition.

* If there is a problem loading the configured container file, include the configured path in the error message.

* Configuration settings for the PSRContainerFakerProvider.

Default values allow users that are not using the provider to not have to set values. These defaults
either match the default values that are already used in constructors or, in the case of the
$phpContainerPath, have a default value that will at least be useful when it is not overridden by
the user.

* Updated documentation.

* Remove the static variables, now that I understand the configuration file.

* Predefine the PsrContainerFakerProviderFactory service with its constructor parameters.

Now that we have default values for all the constructor parameters, we can define the
factory service ourselves. This stops the user from having to say "I'm using this
factory with these parameters" and then having to also say "now give me that service
with those parameters".

* Change namespace of parameters to fakerstan > psrProvider.

* Allow the container path to be null until we absolutely decide that we want to create a PsrContainerFakerProvider.

* Update src/PsrContainerFakerProviderFactory.php

Co-authored-by: Caleb White <cdwhite3@pm.me>

---------

Co-authored-by: Caleb White <cdwhite3@pm.me>